refactor(TALK): SEND MESSAGE - DIR 구조 추가
