+++ src/chat/talk/sendMessage/sendPoto.tsx
@@ -3,7 +3,7 @@ import { Socket } from 'socket.io-client';
 
 import PotoIcon from '../assets/poto.png';
 
-import ChatEvents from '../lib/chatEvents';
+import ChatEvents from '../../lib/chatEvents';
 
 interface SendPotoProps {
     socket: Socket
